chore: 디렉토리 구성 변경
Dae-Hwa/airbnb/#23
