+++ IOS/airbnb/airbnb/Main/ViewController/HomeViewController.swift
@@ -59,7 +59,7 @@ extension HomeViewController: UISearchBarDelegate {
     }
     
     private func pushNextViewController() {
-        let nextStoryBoard = StoryboardFactory.create(.accomodationConditions)
+        let nextStoryBoard = StoryboardFactory.create(.accommodationConditions)
         let popularLocationViewController = ViewControllerFactory.create(from: nextStoryBoard, type: PopularLocationViewController.self)
         self.tabBarController?.navigationItem.backButtonTitle = HomeViewModel.ButtonTitle.back
         self.navigationController?.pushViewController(popularLocationViewController, animated: true)
