+++ iOS/airbnb/airbnb/AccommodationConditions/CalendarSelect/ViewController/CalendarCollectionViewDataSource.swift
@@ -1,0 +1,26 @@
+//
+//  CalendarCollectionViewDataSource.swift
+//  airbnb
+//
+//  Created by Song on 2021/05/27.
+//
+
+import UIKit
+
+class CalendarCollectionViewDataSource: NSObject, UICollectionViewDataSource {
+    
+    func numberOfSections(in collectionView: UICollectionView) -> Int {
+        return 2
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 31
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cellId = CalendarCollectionViewCell.reuseIdentifier
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellId, for: indexPath) as? CalendarCollectionViewCell ?? CalendarCollectionViewCell()
+        return cell
+    }
+
+}
