+++ iOS/airbnb/airbnb/AccommodationConditions/CalendarSelect/ViewModel/CalendarViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  CalendarViewModel.swift
+//  airbnb
+//
+//  Created by Song on 2021/05/27.
+//
+
+import Foundation
