+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/Model/LocationSearchResult.swift
