+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/Model/PopularLocation.swift
