+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/UseCase/SearchResultUseCase.swift
