+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/ViewController/PopularLocationViewController.swift
@@ -31,7 +31,7 @@ final class PopularLocationViewController: UIViewController {
     }
     
     private func setNavigationSearchController() {
-        let storyboard = self.storyboard ?? StoryboardFactory.create(.accomodationConditions)
+        let storyboard = self.storyboard ?? StoryboardFactory.create(.accommodationConditions)
         let searchResultViewController = ViewControllerFactory.create(from: storyboard, type: SearchResultTableViewController.self)
         searchController = LocationSearchController(searchResultsController: searchResultViewController)
         searchResultViewController.delegate = self
@@ -119,7 +119,7 @@ extension PopularLocationViewController: SearchResultDelegate {
     }
     
     private func pushNextViewController(with result: LocationSearchResult) {
-        let storyboard = self.storyboard ?? StoryboardFactory.create(.accomodationConditions)
+        let storyboard = self.storyboard ?? StoryboardFactory.create(.accommodationConditions)
         let nextViewController = ViewControllerFactory.create(from: storyboard, type: CalendarViewController.self)
         nextViewController.location = result
         self.navigationItem.backButtonTitle = PopularLocationViewModel.ButtonTitle.back
