+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/ViewModel/PopularLocationLoadModel.swift
