+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/ViewModel/SearchResultUpdateModel.swift
