+++ iOS/airbnb/airbnb/AccommodationConditions/LocationSearch/ViewModel/SearchResultViewModel.swift
